semana5/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.110.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.29.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.22.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4